User can view a list of students.

+++ tables/students.sql
@@ -4,5 +4,5 @@ CREATE TABLE students (
   id          serial,
   name        text,
   parent_name text,
-  parent_sms  integer
+  parent_sms  bigint
   );
